完善菜单功能

+++ auth-shiro/src/main/java/com/fred/authshiro/model/TbMenu.java
@@ -14,7 +14,7 @@ public class TbMenu implements Serializable {
     /**
      * 父级ID
      */
-    private Long parentId;
+    private Integer parentId;
 
     /**
      * 菜单名称
